pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "g3lobster"
version = "0.1.0"
description = "Google Chat-first named-agent service with persistent local memory"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "google-api-python-client",
    "google-auth-oauthlib",
    "google-auth-httplib2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.21",
    "httpx>=0.25",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["g3lobster*"]

[tool.setuptools.package-data]
g3lobster = [
    "static/index.html",
    "static/css/*.css",
    "static/js/*.js",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[project.scripts]
g3lobster = "g3lobster.main:main"
